homeserver-dashboard/requirements.txt
Flask>=2.2
python-dotenv>=0.21
requests>=2.28
gunicorn>=21.0

